rtt_ros2_params: rename service class, source and header files for consistency
